Move mergeBufferGeometries to geometry.js
